requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector Database
chromadb==0.4.18

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Embeddings & LLM
openai==1.3.5
anthropic==0.7.7

# Text Processing
# langchain==0.0.350  # Temporarily commented out due to dependency conflicts - will use directly when needed
langchain-text-splitters==0.0.1
rank-bm25==0.2.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

